Allow caching Youtube data and using local file
