+++ .gitignore
@@ -57,6 +57,7 @@ typings/
 # gatsby files
 .cache/
 public
+playlist.cache.json
 
 # Mac files
 .DS_Store
