+++ plugins/gatsby-source-playlist/gatsby-node.js
@@ -1,46 +1,4 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const channelMap = require("./channel-map.json");
-const Youtube = require("./youtube").Youtube;
-
-const getAllPlaylistWithVideos = async apiKey => {
-  const yt = new Youtube(apiKey);
-  let allPlaylistWithVideos = [];
-  for (channel of channelMap) {
-    const channelPlaylists = await yt.getChannelPlaylists(channel.channelId);
-    const playlistArr = [];
-    for (playlist of channelPlaylists) {
-      const playlistObj = {
-        id: playlist.id,
-        title: playlist.snippet.title,
-        organisation: playlist.snippet.channelTitle,
-        donationMethod: "<Donation Method>",  // TODO: Pull from some nap.
-        tags: "tags,to,be,implemented",       // TODO: Parse from description?
-        platform: "YouTube",
-        pageUrl: channel.pageUrl,
-        thumbnailUrl: playlist.snippet.thumbnails.medium.url, // Encountered error: thumbnails key (i.e. thumbnails.standard) may not exist.
-      };
-      const playlistVideos = await yt.getPlaylistVideos(playlist.id);
-      const videosArr = playlistVideos.map(video => {
-        return {
-          id: video.snippet.resourceId.videoId,
-          playlistId: video.snippet.playlistId,
-          title: video.snippet.title,
-          asatizah: "<Asatizah name>",        // TODO: Parse from description/title?
-          language: "english",
-          addedOn: video.snippet.publishedAt,
-          videoUrl:
-            "https://www.youtube.com/embed/" +
-            video.snippet.resourceId.videoId,
-        };
-      });
-      playlistObj.videos = videosArr;
-      playlistArr.push(playlistObj);
-    }
-    allPlaylistWithVideos = [...allPlaylistWithVideos, ...playlistArr];
-  }
-  console.log("YouTube API Quota Units used: ", yt.usedQuota);
-  return allPlaylistWithVideos;
-};
+const { getAllPlaylistWithVideos } = require('./utils')
 
 exports.sourceNodes = async (
   { actions, createContentDigest },
@@ -63,7 +21,7 @@ exports.sourceNodes = async (
     return nodeData;
   };
 
-  const playlists = await getAllPlaylistWithVideos(configOptions.apiKey);
+  const playlists = await getAllPlaylistWithVideos(configOptions.apiKey, configOptions.useLocal);
 
   // Creates a new node for each playlist in array.
   playlists.forEach(datum => createNode(processDatum(datum)));
